[v5] [table] promote ColumnHeaderCell2, RowHeaderCell2 APIs
