+++ packages/docs-app/src/examples/table-examples/cellLoadingExample.tsx
@@ -17,7 +17,7 @@ import * as React from "react";
 
 import { RadioGroup } from "@blueprintjs/core";
 import { Example, ExampleProps, handleStringChange } from "@blueprintjs/docs-theme";
-import { Cell, Column, ColumnHeaderCell2, RowHeaderCell2, Table2 } from "@blueprintjs/table";
+import { Cell, Column, ColumnHeaderCell, RowHeaderCell, Table2 } from "@blueprintjs/table";
 
 interface BigSpaceRock {
     [key: string]: number | string;
@@ -122,11 +122,11 @@ export class CellLoadingExample extends React.PureComponent<ExampleProps, CellLo
         const formattedColumnName = columnName
             .replace(/([A-Z])/g, " $1")
             .replace(/^./, firstCharacter => firstCharacter.toUpperCase());
-        return <ColumnHeaderCell2 loading={this.isLoading(0, columnIndex + 1)} name={formattedColumnName} />;
+        return <ColumnHeaderCell loading={this.isLoading(0, columnIndex + 1)} name={formattedColumnName} />;
     };
 
     private renderRowHeaderCell = (rowIndex: number) => {
-        return <RowHeaderCell2 loading={this.isLoading(rowIndex + 1, 0)} name={`${rowIndex + 1}`} />;
+        return <RowHeaderCell loading={this.isLoading(rowIndex + 1, 0)} name={`${rowIndex + 1}`} />;
     };
 
     private isLoading = (rowIndex: number, columnIndex: number) => {
