+++ packages/table/src/deprecatedAliases.ts
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2023 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export {
+    /** @deprecated import ColumnHeaderCell instead */
+    ColumnHeaderCell as ColumnHeaderCell2,
+    /** @deprecated import ColumnHeaderCellProps instead */
+    ColumnHeaderCellProps as ColumnHeaderCellProps2,
+} from "./headers/columnHeaderCell";
+
+export {
+    /** @deprecated import RowHeaderCell instead */
+    RowHeaderCell as RowHeaderCell2,
+    /** @deprecated import RowHeaderCellProps instead */
+    RowHeaderCellProps as RowHeaderCellProps2,
+} from "./headers/rowHeaderCell";
