+++ packages/table/test/columnHeaderCellTests.tsx
@@ -14,19 +14,12 @@
  * limitations under the License.
  */
 
-/**
- * @fileoverview This component is DEPRECATED, and the code is frozen.
- * All changes & bugfixes should be made to ColumnHeaderCell2 instead.
- */
-
-/* eslint-disable deprecation/deprecation, @blueprintjs/no-deprecated-components */
-
 import { expect } from "chai";
-import { shallow } from "enzyme";
+import { mount } from "enzyme";
 import * as React from "react";
-import sinon from "sinon";
+import * as sinon from "sinon";
 
-import { H4, Menu, MenuItem } from "@blueprintjs/core";
+import { Classes as CoreClasses, H4, Menu, MenuItem } from "@blueprintjs/core";
 
 import { ColumnHeaderCell, ColumnHeaderCellProps } from "../src";
 import * as Classes from "../src/common/classes";
@@ -59,14 +52,20 @@ describe("<ColumnHeaderCell>", () => {
 
     it("passes index prop to nameRenderer callback if index was provided", () => {
         const renderNameStub = sinon.stub();
-        renderNameStub.returns("string");
+        renderNameStub.returns(<span>name</span>);
         const NAME = "my-name";
         const INDEX = 17;
-        shallow(<ColumnHeaderCell index={INDEX} name={NAME} nameRenderer={renderNameStub} />);
+        mount(<ColumnHeaderCell index={INDEX} name={NAME} nameRenderer={renderNameStub} />);
         expect(renderNameStub.firstCall.args).to.deep.equal([NAME, INDEX]);
     });
 
     describe("Custom renderer", () => {
+        const menuClickSpy = sinon.spy();
+
+        beforeEach(() => {
+            menuClickSpy.resetHistory();
+        });
+
         it("renders custom name", () => {
             const columnHeaderCellRenderer = (columnIndex: number) => {
                 return <ColumnHeaderCell name={`COLUMN-${columnIndex}`} />;
@@ -89,28 +88,36 @@ describe("<ColumnHeaderCell>", () => {
             expect(text).to.equal("Header of 2");
         });
 
-        it("renders custom menu items", () => {
-            const menuClickSpy = sinon.spy();
-            const menu = renderMenu(menuClickSpy);
-            const renderMenuFn = () => menu;
-
-            const columnHeaderCellRenderer = (columnIndex: number) => {
-                return <ColumnHeaderCell name={`COL-${columnIndex}`} menuRenderer={renderMenuFn} />;
-            };
-            const table = harness.mount(createTableOfSize(3, 2, { columnHeaderCellRenderer }));
-            expectMenuToOpen(table, menuClickSpy);
-        });
-
         it("renders custom menu items with a menuRenderer callback", () => {
-            const menuClickSpy = sinon.spy();
-            const menu = renderMenu(menuClickSpy);
-            const menuRenderer = sinon.stub().returns(menu);
-
             const columnHeaderCellRenderer = (columnIndex: number) => (
-                <ColumnHeaderCell name={`COL-${columnIndex}`} menuRenderer={menuRenderer} />
+                <ColumnHeaderCell name={`COL-${columnIndex}`} menuRenderer={renderMenu} />
             );
             const table = harness.mount(createTableOfSize(3, 2, { columnHeaderCellRenderer }));
-            expectMenuToOpen(table, menuClickSpy);
+            expectMenuToOpen(table);
+
+            // attempt to click one of the menu items
+            ElementHarness.document().find('[data-icon="export"]')!.mouse("click");
+            expect(menuClickSpy.called).to.be.true;
+        });
+
+        it("custom menu supports popover props", () => {
+            const expectedMenuPopoverProps = {
+                placement: "top" as const,
+                popoverClassName: "test-popover-class",
+            };
+            const columnHeaderCellRenderer = (columnIndex: number) => (
+                <ColumnHeaderCell
+                    name={`COL-${columnIndex}`}
+                    menuRenderer={renderMenu}
+                    menuPopoverProps={expectedMenuPopoverProps}
+                />
+            );
+            const table = harness.mount(createTableOfSize(3, 2, { columnHeaderCellRenderer }));
+            expectMenuToOpen(table);
+            const popover = ElementHarness.document().find(`.${CoreClasses.POPOVER}`);
+            expect(popover.hasClass(expectedMenuPopoverProps.popoverClassName)).to.be.true;
+            expect(popover.hasClass(`${CoreClasses.POPOVER_CONTENT_PLACEMENT}-${expectedMenuPopoverProps.placement}`))
+                .to.be.true;
         });
 
         it("renders loading state properly", () => {
@@ -124,49 +131,46 @@ describe("<ColumnHeaderCell>", () => {
             );
         });
 
-        function renderMenu(menuClickSpy: sinon.SinonSpy) {
+        function renderMenu() {
             return (
                 <Menu>
-                    <MenuItem className="export-item" icon="export" onClick={menuClickSpy} text="Teleport" />
+                    <MenuItem icon="export" onClick={menuClickSpy} text="Teleport" />
                     <MenuItem icon="sort-alphabetical-desc" onClick={menuClickSpy} text="Down with ZA!" />
                     <MenuItem icon="curved-range-chart" onClick={menuClickSpy} text="Psi" />
                 </Menu>
             );
         }
 
-        function expectMenuToOpen(table: ElementHarness, menuClickSpy: sinon.SinonSpy) {
+        function expectMenuToOpen(table: ElementHarness) {
             table.find(`.${Classes.TABLE_COLUMN_HEADERS}`)!.mouse("mousemove");
-            // menu element is the popover target
-            table.find(`.${Classes.TABLE_TH_MENU}`).mouse("click");
-            ElementHarness.document().find(".export-item").mouse("click");
-            expect(menuClickSpy.called).to.be.true;
+            const target = table.find(`.${Classes.TABLE_TH_MENU}.${CoreClasses.POPOVER_TARGET}`)!;
+            target.mouse("click");
+            expect(target.hasClass(CoreClasses.POPOVER_OPEN)).to.be.true;
         }
     });
 
-    // TODO: re-enable these tests when we switch to enzyme's testing harness instead of our own,
-    // so that we can supply a react context with enableColumnInteractionBar: true
-    // see https://github.com/palantir/blueprint/issues/2076
-    describe.skip("Reorder handle", () => {
+    describe("Reorder handle", () => {
         const REORDER_HANDLE_CLASS = Classes.TABLE_REORDER_HANDLE_TARGET;
 
         it("shows reorder handle in interaction bar if reordering and interaction bar are enabled", () => {
-            const element = mount({ enableColumnReordering: true });
-            expect(element.find(`.${Classes.TABLE_INTERACTION_BAR} .${REORDER_HANDLE_CLASS}`)!.exists()).to.be.true;
+            const wrapper = mountHeaderCell();
+            expect(wrapper.find(`.${Classes.TABLE_INTERACTION_BAR} .${REORDER_HANDLE_CLASS}`)!.exists()).to.be.true;
         });
 
         it("shows reorder handle next to column name if reordering enabled but interaction bar disabled", () => {
-            const element = mount({ enableColumnReordering: true });
-            expect(element.find(`.${Classes.TABLE_COLUMN_NAME} .${REORDER_HANDLE_CLASS}`)!.exists()).to.be.true;
+            const wrapper = mountHeaderCell({ enableColumnInteractionBar: false });
+            expect(wrapper.find(`.${Classes.TABLE_COLUMN_NAME} .${REORDER_HANDLE_CLASS}`)!.exists()).to.be.true;
         });
 
-        function mount(props: Partial<ColumnHeaderCellProps>) {
-            const element = harness.mount(
+        function mountHeaderCell(props?: Partial<ColumnHeaderCellProps>) {
+            return mount(
                 <ColumnHeaderCell
-                    enableColumnReordering={props.enableColumnReordering}
+                    enableColumnInteractionBar={true}
                     reorderHandle={<div className={REORDER_HANDLE_CLASS} />}
+                    enableColumnReordering={true}
+                    {...props}
                 />,
             );
-            return element;
         }
     });
 });
