+++ packages/table/test/index.ts
@@ -20,7 +20,6 @@ import "./batcherTests.tsx";
 import "./cellTests.tsx";
 import "./clipboardTests.ts";
 import "./columnHeaderCellTests.tsx";
-import "./columnHeaderCell2Tests.tsx";
 import "./columnTests.tsx";
 import "./common/internal";
 import "./editableCell2Tests.tsx";
@@ -40,7 +39,6 @@ import "./rectTests.ts";
 import "./regionsTests.ts";
 import "./reorderableTests.tsx";
 import "./resizableTests.tsx";
-import "./rowHeaderCell2Tests.tsx";
 import "./rowHeaderCellTests.tsx";
 import "./selectableTests.tsx";
 import "./selectionTests.tsx";
